Improve and simplify startup code
Move unrelated updates/version code to an extension on ICommandApp.
