+++ src/OpenLaw/OpenLaw.csproj
@@ -26,11 +26,7 @@
     <PackageReference Include="Polly" Version="8.5.2" />
     <PackageReference Include="Spectre.Console.Cli" Version="0.49.1" />
     <PackageReference Include="Spectre.Console.Json" Version="0.49.1" />
-    <PackageReference Include="ThisAssembly.Strings" Version="2.0.12" PrivateAssets="all" />
-    <PackageReference Include="ThisAssembly.AssemblyInfo" Version="2.0.12" PrivateAssets="all" />
-    <PackageReference Include="ThisAssembly.Project" Version="2.0.12" PrivateAssets="all" />
-    <PackageReference Include="ThisAssembly.Constants" Version="2.0.12" PrivateAssets="all" />
-    <PackageReference Include="ThisAssembly.Resources" Version="2.0.12" PrivateAssets="all" />
+    <PackageReference Include="ThisAssembly" Version="2.0.12" PrivateAssets="all" />
     <PackageReference Include="YamlDotNet" Version="16.3.0" />
     <PackageReference Include="Markdown2Pdf" Version="2.2.2" />
   </ItemGroup>
