+++ src/dotnet-openlaw/dotnet-openlaw.csproj
@@ -22,15 +22,6 @@
   </ItemGroup>
 
   <ItemGroup>
-    <ProjectProperty Include="ToolCommandName" />
-    <ProjectProperty Include="BuildDate" />
-    <ProjectProperty Include="BuildRef" />
-    <ProjectProperty Include="PackageId" />
-    <ProjectProperty Include="PackageVersion" />
-    <ProjectProperty Include="SLEET_FEED_URL" />
-  </ItemGroup>
-
-  <ItemGroup>
     <CommandHelp Include="sync" />
     <CommandHelp Include="convert" />
     <CommandHelp Include="format" />
